Tune SQS consumer behavior by setting a wait timeout of 1s.
Setting to zero (the previous behavior) fails to wait long enough for new messages to appear;
this causes long delays on message delivery. Setting to larger values causes the daemon to block
waiting on SQS; this interferes with daemon interruption and shutdown.

At the same time: add the simple daemon from the README to the examples and modify the example
message to better enable diagnostics.