Enable fluent decorators for registering PubSub schemas and handlers.
This is best demonstrated by example:

    @schema
    class MySchema(PubSubMessageSchema):
        my_field = fields.String()

    @binding("my_handler")
    @handles(MySchema)
    class MyHandler(object):

        def __init__(self, graph):
            self.collaborator = graph.collaborator

        def __call__(self, message):
            self.collaborator.do_something(message)
            return True

The underlying change involves moving the schema and handler mappings into a centralized,
singleton registry. This is a small loss in terms of object graph purity -- these mappings
are stored in singletons, not the object graph because it's rather inconvenient to enforce
that the graph exists at the point in time that these types are declared -- but the benefit
is a great deal more clarity to what was a frequent source of usage errors.

The legacy registry mechanism (namely, dictionaries bound to the graph) still exists, but
is implemented by loading these dictionaries into the registry when the graph binds the
registries. This behavior still needs some testing in anger.