Create decorator to allow for automatic message deferment
This decorator is intended to be used at the controller configuration
level, with the idea being that any messages produced within the context of a
controller function will be produced after it has completed.

This will allow these messages to be published after the resource
they're referencing has been persisted, guaranteeing that it will be available
when fetched by other message handlers.

Note that this is not a full solution; it's not uncommon for components
to call other components directly, and this decorator will not change those
references in place. As is, it will only apply the messages produced within a
function by the specified component.