Fix case where opaque_data contains message ID
When we publish messages, we tend to publish them with the entirety of
the opaque data, which itself includes the previous message ID of the
message processed by another producer.

This reverts back to the original, previous behavior where we start with
the opaque_data, and override values as they're specified by
message_id/kwargs.

Note that there may be other cases of repeated kwargs. This only takes
message_id into account for now.