Move processing limit logic to dispatcher
The logic to maybe stop processing a message didn't naturally belong in the backoff policy (which decides what to do on failed processing). Instead it make more sense for it to live pre-message processing, in the SQSDispatcher.
It makes this logic seemingly redundant with the X-Request-Ttl validation. However, a message's TTL can't be changed by the handler because messages are immutable, meaning that we can't use the X-Request-Ttl to count processing iterations.