Add metrics per-handled message batch
Adds two new metrics: the time needed to process a batch of messages,
and the count of relevant messages per batch.

The previous metrics, while useful for per-message handling statistics,
weren't useful enough when trying to analyze message-handling from the
consumer's viewpoint. For example, it would be difficult to see how
long, under load, it took for any given service to handle a batch of
messages.