Release/glob 33011 changed messages (#168) (#169)
* GLOB-33011 - changed schema has two new fields. handler checks them (#167)

Why?
When publishing a changed(Class) handler could not determine if it caught the message
before the change was committed to the DB. To resolve this, we're adding
a new default message for changed events. message has `field_name` and `new_value`
on it. Handler checks that the retrieved object has the correct value, and in
case it isn't, nacks and retries later.

What I've done:
1. added a new `ChangedMessageSchema` type with `field_name` and `new_value`
2. that type is the default for changed events
3. amended `URIHandler` to check the new value and nack if doesn't fit the message

* bump version to 1.26.0