Syntactic sugar to give names to function argument and result
We don't have an easy, explicity way to temporarily change the context for a single function call.
This limits the amount of code reuse we can do: currently we'd need to write two functions with different argument names, calling the same underlying logic.
This new syntax allows us to do that much more easily.