Remove media type for convention-driven message schema
**Why?**
A media type was being passed in to our convention-driven message schemas. This was a hacky way of making them work with the pubsub registry; this has since been fixed by allowing the registry to store media types without an associated schema, in which case it defaults back to convention.

Having a required constructor argument for a schema is bad practice in my opinion, especially since that argument is not actually needed to define the schema.

**What?**
Remove `media_type` from constructor for `URIMessageSchema` and `IdentityMessageSchema`.