Expose timeout configuration for sns client connections
Enables the sns client to have read/connect timeouts to be configured,
such that it isn't limited by the default 60s limit.

We've observed (in real usage) that connections may end up taking an
entire 60s, timing out, and then succeeding as expected on the next
attempt. This allows one to experiment with lowering (or raising) that
limit as needed per-application.