Instantiate schemas with unknown=EXCLUDE
**Why?**
Validating schemas on serialization using Marshmallow presents some issues, due to the fact that it is not in line with Marshmallow's philosophy (see this [GitHub issue](https://github.com/marshmallow-code/marshmallow/issues/682)). The main issue is that a schema's `validate` method ignores `dump_only` fields, so those are treated as unknown fields.
However, validation is still useful in highlighting fields that should be marked as nullable in the schema.

**What?**
- Bring back schema validation on serialization.
- Use `unknown=EXCLUDE` on schema instantiation, meaning that unknown fields won't trigger errors.